+++ syntax/specialization/src/test/scala/org/enso/syntax/text/Parser.scala
@@ -235,30 +235,30 @@ class ParserSpec extends FlatSpec with Matchers {
 
   //// Valid ////
 
-  "()"                 ?= "(" II ")"
-  "( )"                ?= "(" I_I ")"
-  "( (  )   )"         ?= "(" I_ ("(" I__I ")") I___ ")"
-  "(a)"                ?= "(" I "a" I ")"
-  "((a))"              ?= "(" I ("(" I "a" I ")") I ")"
-  "(((a)))"            ?= "(" I ("(" I ("(" I "a" I ")") I ")") I ")"
-  "( (  a   )    )"    ?= "(" I_ ("(" I__ "a" I___ ")") I____ ")"
-  "if a  then   b"     ?= "if" I1_ "a" I1__ "then" I1___ "b"
-  "if a then b else c" ?= "if" I1_ "a" I1_ "then" I1_ "b" I1_ "else" I1_ "c"
-  "(if a then  b) c"   ?= "(" I ("if" I1_ "a" I1_ "then" I1__ "b") I ")" $_ "c"
-  "a (b c) d"          ?= "a" $_ ("(" I ("b" $_ "c") I ")") $_ "d"
-
-  "(if a then b) else c" ?=
-  "(" I ("if" I1_ "a" I1_ "then" I1_ "b") I ")" $_ "else" $_ "c"
-
-  //// Invalid ////
-
-  val _then_else = List(List("then"), List("then", "else"))
-
-  "("           ?= "(" Ix ")"
-  "(("          ?= "(" I ("(" Ix ")") Ix ")"
-  "if"          ?= "if" Ixx (_then_else: _*)
-  "(if a) then" ?= "(" I ("if" I_ "a" Ixx (_then_else: _*)) I ")" $_ "then"
-  "if (a then)" ?= "if" I_ ("(" I ("a" $_ "then") I ")") Ixx (_then_else: _*)
+//  "()"                 ?= "(" II ")"
+//  "( )"                ?= "(" I_I ")"
+//  "( (  )   )"         ?= "(" I_ ("(" I__I ")") I___ ")"
+//  "(a)"                ?= "(" I "a" I ")"
+//  "((a))"              ?= "(" I ("(" I "a" I ")") I ")"
+//  "(((a)))"            ?= "(" I ("(" I ("(" I "a" I ")") I ")") I ")"
+//  "( (  a   )    )"    ?= "(" I_ ("(" I__ "a" I___ ")") I____ ")"
+//  "if a  then   b"     ?= "if" I1_ "a" I1__ "then" I1___ "b"
+//  "if a then b else c" ?= "if" I1_ "a" I1_ "then" I1_ "b" I1_ "else" I1_ "c"
+//  "(if a then  b) c"   ?= "(" I ("if" I1_ "a" I1_ "then" I1__ "b") I ")" $_ "c"
+//  "a (b c) d"          ?= "a" $_ ("(" I ("b" $_ "c") I ")") $_ "d"
+//
+//  "(if a then b) else c" ?=
+//  "(" I ("if" I1_ "a" I1_ "then" I1_ "b") I ")" $_ "else" $_ "c"
+//
+//  //// Invalid ////
+//
+//  val _then_else = List(List("then"), List("then", "else"))
+//
+//  "("           ?= "(" Ix ")"
+//  "(("          ?= "(" I ("(" Ix ")") Ix ")"
+//  "if"          ?= "if" Ixx (_then_else: _*)
+//  "(if a) then" ?= "(" I ("if" I_ "a" Ixx (_then_else: _*)) I ")" $_ "then"
+//  "if (a then)" ?= "if" I_ ("(" I ("a" $_ "then") I ")") Ixx (_then_else: _*)
 
   //  "import Std.Math" ?= "foo"
